// inkreta5/src/app/components/customer/invoicelist/models/invoicelist.model.ts
export interface InvoiceList {
  invoice_status: string;
  total_invoice_value:number;
  invoiceduedate: string;
  invoicedate: string;
  creditdays: number;
  invoicenum: string;
  status:string;
  supply_type:string;
  document_ref_id: string;
  customer_partner_id:string;
  company_name:string;
  results:[]
}